Hide and Show all notes functionality, and Setting's popover
Nicole was right about how we only need to set the hidden variable in
our text labels (very convenient!). I gave QuickStageView a variable to
keep track of whether or not the notes are being shown.

Two new files: SettingsView.h and SettingsView.m are new files which
replaces the previous DimensionsView.h and DimensionsView.m. It
currently just shows a table of with no actual functionality but to
show alerts about their supposed functionality.