Fixed Pan Gesture Control and added Pinch and Rotate Gesture Controls
Note to Nicole: (totally unrelated to this commit) You might want to
read my comment in SetPieceListView.m (command+f : "JNN")

A lot of changes in TVGestureController to support the new gesture
controls. In Actor, Note, and SetPiece, there is a new
CGAffineTransform variable called scaleRotationMatrix which stores the
scale and rotation for the pieces.