Almost got undo movement
Can undo a note movement. One problem when starting to undo a chain of
actions. It will first undo the current change, which should be in the
same position, but it causes the note to shift down a bit. Then it will
start going backwards and move to the previous position.